+++ src/hera/cron_workflow.py
@@ -17,6 +17,7 @@ from argo_workflows.models import (
 
 from hera.cron_workflow_service import CronWorkflowService
 from hera.task import Task
+from hera.volumes import Volume
 
 
 class CronWorkflow:
@@ -75,7 +76,9 @@ class CronWorkflow:
             dag=self.dag_template,
             parallelism=self.parallelism,
         )
-        self.spec = IoArgoprojWorkflowV1alpha1WorkflowSpec(templates=[self.template], entrypoint=self.name)
+        self.spec = IoArgoprojWorkflowV1alpha1WorkflowSpec(
+            templates=[self.template], entrypoint=self.name, volumes=[], volume_claim_templates=[]
+        )
         if self.service_account_name:
             setattr(self.template, 'service_account_name', self.service_account_name)
             setattr(self.spec, 'service_account_name', self.service_account_name)
@@ -105,35 +108,16 @@ class CronWorkflow:
         if not all(ts):
             return
 
-        if not hasattr(self.spec, 'volume_claim_templates'):
-            setattr(self.spec, 'volume_claim_templates', [])
-
         for t in ts:
             self.spec.templates.append(t.argo_template)
 
-            if t.resources.volume:
-                if hasattr(self.spec, 'volume_claim_template'):
-                    self.spec.volume_claim_templates.append(t.resources.volume.get_claim_spec())
-                else:
-                    setattr(self.spec, 'volume_claim_templates', [t.resources.volume.get_claim_spec()])
-
-            if t.resources.existing_volume:
-                if hasattr(self.spec, 'volumes'):
-                    self.spec.volumes.append(t.resources.existing_volume.get_volume())
-                else:
-                    setattr(self.spec, 'volumes', [t.resources.existing_volume.get_volume()])
-
-            if t.resources.empty_dir_volume:
-                if hasattr(self.spec, 'volumes'):
-                    self.spec.volumes.append(t.resources.empty_dir_volume.get_volume())
-                else:
-                    setattr(self.spec, 'volumes', [t.resources.empty_dir_volume.get_volume()])
-
-            if t.resources.secret_volume:
-                if hasattr(self.spec, 'volumes'):
-                    self.spec.volumes.append(t.resources.secret_volume.get_volume())
-                else:
-                    setattr(self.spec, 'volumes', [t.resources.secret_volume.get_volume()])
+            if t.resources.volumes:
+                for vol in t.resources.volumes:
+                    if isinstance(vol, Volume):
+                        # dynamically provisioned volumes need associated claims on the workflow spec
+                        self.spec.volume_claim_templates.append(vol.get_claim_spec())
+                    else:
+                        self.spec.volumes.append(vol.get_volume())
 
             self.dag_template.tasks.append(t.argo_task)