+++ tests/test_cron_workflow.py
@@ -43,7 +43,7 @@ def test_cwf_adds_task_volume(cw, no_op):
     t = Task(
         't',
         no_op,
-        resources=Resources(volume=Volume(name='v', size='1Gi', mount_path='/', storage_class_name='custom')),
+        resources=Resources(volumes=[Volume(name='v', size='1Gi', mount_path='/', storage_class_name='custom')]),
     )
     cw.add_task(t)
 
@@ -55,7 +55,7 @@ def test_cwf_adds_task_volume(cw, no_op):
 
 
 def test_wf_adds_task_secret_volume(cw, no_op):
-    t = Task('t', no_op, resources=Resources(secret_volume=SecretVolume(name='s', secret_name='sn', mount_path='/')))
+    t = Task('t', no_op, resources=Resources(volumes=[SecretVolume(name='s', secret_name='sn', mount_path='/')]))
     cw.add_task(t)
 
     vol = cw.spec.volumes[0]
@@ -64,7 +64,7 @@ def test_wf_adds_task_secret_volume(cw, no_op):
 
 
 def test_cwf_adds_task_existing_checkpoints_staging_volume(cw, no_op):
-    t = Task('t', no_op, resources=Resources(existing_volume=ExistingVolume(name='v', mount_path='/')))
+    t = Task('t', no_op, resources=Resources(volumes=[ExistingVolume(name='v', mount_path='/')]))
     cw.add_task(t)
 
     vol = cw.spec.volumes[0]
@@ -76,7 +76,7 @@ def test_cwf_adds_task_existing_checkpoints_prod_volume(cw, no_op):
     t = Task(
         't',
         no_op,
-        resources=Resources(existing_volume=ExistingVolume(name='vol', mount_path='/')),
+        resources=Resources(volumes=[ExistingVolume(name='vol', mount_path='/')]),
     )
     cw.add_task(t)
 
@@ -86,7 +86,7 @@ def test_cwf_adds_task_existing_checkpoints_prod_volume(cw, no_op):
 
 
 def test_cwf_adds_task_empty_dir_volume(cw, no_op):
-    t = Task('t', no_op, resources=Resources(empty_dir_volume=EmptyDirVolume(name='v')))
+    t = Task('t', no_op, resources=Resources(volumes=[EmptyDirVolume(name='v')]))
     cw.add_task(t)
 
     vol = cw.spec.volumes[0]