+++ tests/test_resources.py
@@ -21,7 +21,7 @@ def test_init_raises_on_invalid_cpu():
 
 def test_init_volume_error_propagates():
     with pytest.raises(ValidationError):
-        Resources(volume=Volume(size='1', mount_path='/path'))
+        Resources(volumes=[Volume(size='1', mount_path='/path')])
 
 
 def test_init_passes():
@@ -31,19 +31,33 @@ def test_init_passes():
         min_mem='2Gi',
         max_mem='3Gi',
         gpus=1,
-        volume=Volume(size='10Gi', mount_path='/path'),
-        existing_volume=ExistingVolume(name='test', mount_path='/path2'),
-        secret_volume=SecretVolume(name='secret', secret_name='secret_name', mount_path="/path3"),
+        volumes=[
+            Volume(size='10Gi', mount_path='/path'),
+            ExistingVolume(name='test', mount_path='/path2'),
+            SecretVolume(name='secret', secret_name='secret_name', mount_path="/path3"),
+        ],
     )
     assert r.min_cpu == 1
     assert r.max_cpu == 2
     assert r.min_mem == '2Gi'
     assert r.max_mem == '3Gi'
     assert r.gpus == 1
-    assert r.volume.size == '10Gi'
-    assert r.volume.mount_path == '/path'
-    assert r.existing_volume.mount_path == '/path2'
-    assert r.secret_volume.mount_path == '/path3'
+
+    vol = r.volumes[0]
+    assert isinstance(vol, Volume)
+    assert vol.size == '10Gi'
+    assert vol.mount_path == '/path'
+
+    ex_vol = r.volumes[1]
+    assert isinstance(ex_vol, ExistingVolume)
+    assert ex_vol.name == 'test'
+    assert ex_vol.mount_path == '/path2'
+
+    sc_vol = r.volumes[2]
+    assert isinstance(sc_vol, SecretVolume)
+    assert sc_vol.name == 'secret'
+    assert sc_vol.secret_name == 'secret_name'
+    assert sc_vol.mount_path == '/path3'
 
 
 def test_max_set_to_min_if_max_not_specified_with_overwrite():
@@ -63,5 +77,5 @@ def test_max_not_set_to_min_if_max_not_specified_with_no_overwrite():
 
 
 def test_secret_volume_name_generated_when_not_specified():
-    r = Resources(secret_volume=SecretVolume(secret_name="sn", mount_path="/path"))
-    assert r.secret_volume.name
+    s = SecretVolume(secret_name="sn", mount_path="/path")
+    assert s.name is not None and s.name is not ""